# openapi.yaml
openapi: 3.0.1
info:
  title: eCMR - API
  description: eCMR - API
  contact: {
    "name": "Open Logistics Foundation\n",
    "url": "https://openlogisticsfoundation.org/",
    "email": "ecmr-feedback@lists.openlogisticsfoundation.org"
  }
  version: v0.2.0
servers:
- url: http://localhost:8081/api
  description: Generated server url
security:
- OAUth2: []
...
paths:
...
  /document:
    get:
      tags:
        - Document
      operationId: getEcmrDocuments
      parameters:
        - name: ecmrId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentModel"
    post:
      tags:
        - Document
      operationId: uploadDocumentToEcmr
      parameters:
        - name: ecmrId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: OK
  /document/{id}/download:
    get:
      tags:
        - Document
      operationId: downloadDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: binary
  /document/{id}:
    delete:
      tags:
        - Document
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    DocumentModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        uploadDate:
          type: string
          format: date-time
        mimeType:
          type: string
        size:
          type: integer
          format: int32
...
  securitySchemes:
    OAUth2:
      type: oauth2
      flows:
        authorizationCode: {}